# sistema-importacion/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Importación API",
        "description": "API REST y SOAP para gestión de importaciones",
        "contact": {
            "email": "admin@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logout exitoso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/proveedores": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Listar todos los proveedores",
                "operationId": "7f6ed20dc86f1682aa7e31d71afa37df",
                "responses": {
                    "200": {
                        "description": "Lista de proveedores"
                    }
                }
            },
            "post": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Crear un nuevo proveedor",
                "operationId": "a1046f3d7a3c804651bf420761995e57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "pais",
                                    "email"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "pais": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor creado exitosamente"
                    }
                }
            }
        },
        "/api/proveedores/{id}": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Obtener un proveedor específico",
                "operationId": "5aa647e1f6ead15a2c088ecc1fbb12bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor obtenido exitosamente"
                    }
                }
            },
            "put": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Actualizar un proveedor",
                "operationId": "28ff17fc2a31c61745cbe9a18da5aa32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor actualizado exitosamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Eliminar un proveedor",
                "operationId": "2b9dc197602ee96f15138b4ad2024fab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor eliminado exitosamente"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Operaciones de autenticación"
        },
        {
            "name": "Proveedores",
            "description": "Operaciones CRUD para proveedores"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ]
}